+++ engine/js-transformer.js
@@ -1,0 +1,27 @@
+const renderFile = require('./render-file');
+
+class Engine {
+  constructor(files, metalsmith, options) {
+    this.files = files;
+    this.options = options;
+    this.metadata = metalsmith.metadata();
+    this.source = metalsmith.source();
+  }
+
+  render(filename) {
+    return new Promise((resolve) => {
+      // Transform file
+      const newname = renderFile(this.files, filename, this.metadata, this.source, this.options);
+
+      // If there were transforms store new file and delete the old
+      if (newname !== filename) {
+        this.files[newname] = this.files[filename];
+        delete this.files[filename];
+      }
+
+      resolve();
+    });
+  }
+}
+
+module.exports = Engine;
