+++ lib/index.js
@@ -1,34 +1,32 @@
 'use strict';
 
-const debug = require('debug')('metalsmith-in-place');
-const JsTransformer = require('./js-transformer');
+const JsTransformer = require('../engine/js-transformer'); // will be moved to an external module
+const match = require('multimatch');
 
-/**
- * Local
- */
+// Checks if files match the provided pattern
+function check(filename, pattern) {
+  return match(filename, pattern)[0];
+}
 
-const check = require('./check');
-
-/**
- * Main export
- */
-
-module.exports = function plugin(opts) {
-  opts = opts || {};
-
-  // Parse options and set defaults
-  const Engine = opts.engine || JsTransformer;
-  const engineOptions = opts.engineOptions || {};
-  const pattern = opts.pattern || '**';
+function plugin(options = {}) {
+  // Defaults
+  const pattern = options.pattern || '**';
+  const Engine = options.engine || JsTransformer;
+  const engineOptions = options.engineOptions || {};
 
   return function (files, metalsmith, done) {
     const engine = new Engine(files, metalsmith, engineOptions);
 
-    Object.keys(files).filter((file) => check(file, pattern)).forEach((filename) => {
-      engine.render(filename);
-    });
+    // Map all files that should be processed to an array of promises
+    const promises = Object.keys(files)
+      .filter(filename => check(filename, pattern))
+      .map(filename => engine.render(filename));
 
-    debug('all files processed');
-    done();
+    // Call done callback when all promises are resolved
+    Promise.all(promises)
+      .then(() => done())
+      .catch(error => done(error));
   };
-};
+}
+
+module.exports = plugin;
